perf: apply optimization.md guidelines across katgpt-core and src
katgpt-core:
- peira.rs: SIMD-accelerate Cholesky inner dot products via simd_dot_f64
  (~2-4x speedup on f64 accumulations for k=512)
- questbench.rs: narrower bool→f32 cast (u8 intermediate → better AVX2 IR)
- spectral_hierarchy.rs: idiomatic collect() instead of with_capacity+extend
- simd.rs: branch-free max in attention scoring loops
- attention.rs, parallax_attn.rs: branch-free max in scoring loops

src:
- transformer.rs, turboquant/forward.rs, octopus/forward.rs,
  spectralquant/forward.rs, sp_kv/forward.rs, speculative/drafter_lora.rs:
  branch-free max_score = max_score.max(score) in QK scoring loops
- spectralquant/spectral_kv_cache.rs: eliminate Vec<Vec<f32>> allocation
  in rayon parallel dequantize — single flat buffer + par_chunks_mut
- spectralquant/types.rs: reorder struct fields by alignment convention
- turboquant/kv_cache.rs: #[inline] on hot-path helpers
  (mat_vec_into, mat_vec_t_into, packed_len, pack/unpack_indices_into)
- newton_schulz.rs: branch-free max and cleanup

+++ crates/katgpt-core/src/questbench.rs
@@ -563,21 +563,21 @@ fn score_relevance_into(
         &mut valid[..limit],
     );
     // Chunked bool→f32 conversion (8 at a time) for AVX2 auto-vectorization.
-    // Uses `bool as usize` (0 or 1) for branch-free conversion.
+    // Uses `f32::from(bool as u8)` for branch-free conversion (0 or 1).
     let mut i = 0;
     while i + 8 <= limit {
-        buf[i] = valid[i] as usize as f32;
-        buf[i + 1] = valid[i + 1] as usize as f32;
-        buf[i + 2] = valid[i + 2] as usize as f32;
-        buf[i + 3] = valid[i + 3] as usize as f32;
-        buf[i + 4] = valid[i + 4] as usize as f32;
-        buf[i + 5] = valid[i + 5] as usize as f32;
-        buf[i + 6] = valid[i + 6] as usize as f32;
-        buf[i + 7] = valid[i + 7] as usize as f32;
+        buf[i] = valid[i] as u8 as f32;
+        buf[i + 1] = valid[i + 1] as u8 as f32;
+        buf[i + 2] = valid[i + 2] as u8 as f32;
+        buf[i + 3] = valid[i + 3] as u8 as f32;
+        buf[i + 4] = valid[i + 4] as u8 as f32;
+        buf[i + 5] = valid[i + 5] as u8 as f32;
+        buf[i + 6] = valid[i + 6] as u8 as f32;
+        buf[i + 7] = valid[i + 7] as u8 as f32;
         i += 8;
     }
     for j in i..limit {
-        buf[j] = valid[j] as usize as f32;
+        buf[j] = valid[j] as u8 as f32;
     }
     buf[limit..].fill(0.0);
 }
